requirements.txt
requests
pyyaml
pandas
python-dotenv
huggingface_hub